General code cleanups
* Removed some unused includes and forward declarations.

* Use std::unique_ptr to automate cleanup.

* Use TextRenderer::renderText in BrowserBox to avoid code duplication.

* Removed unused STATE_NORMAL from StateFlags.

* Small layout fix in ServerDialog.

* Reduced rewrapping delay in BrowserBox to 33ms and disabled it
  entirely when there are no more than 1000 lines to rewrap. The
  rewrapping is usually fast enough.

+++ src/gui/charselectdialog.cpp
@@ -150,7 +150,7 @@ CharSelectDialog::CharSelectDialog(LoginData *loginData):
     for (int i = 0; i < (int)mLoginData->characterSlots; i++)
     {
         mCharacterEntries.push_back(new CharacterDisplay(this));
-        place(i % SLOTS_PER_ROW, (int)i / SLOTS_PER_ROW, mCharacterEntries[i]);
+        place(i % SLOTS_PER_ROW, i / SLOTS_PER_ROW, mCharacterEntries[i]);
     }
 
     reflowLayout();
